+++ resources/templates/resultsDisplayClient.js
@@ -1,4 +1,4 @@
-import {jml, $} from 'jamilih';
+import {jml, $, body} from 'jamilih';
 import Templates from './index.js';
 import {dialogs} from '../utils/dialogs.js';
 
@@ -22,6 +22,6 @@ export default {
                 dialogs.alert(err.message);
             }
         }
-        jml(...html, document.body);
+        jml(...html, body);
     }
 };